Entity Cart & Entity CartItem

+++ src/Entity/Product.php
@@ -25,9 +25,9 @@ class Product
     #[Groups('products', 'product')]
     private ?string $description = null;
 
-    #[ORM\Column]
+    #[ORM\Column(type: "float")]
     #[Groups('products', 'product')]
-    private ?int $price = null;
+    private ?float $price = null;
 
     /**
      * @var Collection<int, Picture>
@@ -73,12 +73,12 @@ class Product
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
 
-    public function setPrice(int $price): static
+    public function setPrice(float $price): static
     {
         $this->price = $price;
         return $this;
